Develop

+++ herd-code/herd-dao/src/test/java/org/finra/herd/dao/BusinessObjectDefinitionDaoTest.java
@@ -152,7 +152,7 @@ public class BusinessObjectDefinitionDaoTest extends AbstractDaoTest
         List<BusinessObjectDefinitionEntity> businessObjectDefinitionEntities =
             businessObjectDefinitionDaoTestHelper.createExpectedBusinessObjectDefinitionEntities();
 
-        List<Integer> businessObjectDefinitionIds = new ArrayList<>();
+        List<Long> businessObjectDefinitionIds = new ArrayList<>();
 
         businessObjectDefinitionEntities.forEach(businessObjectDefinitionEntity -> businessObjectDefinitionIds.add(businessObjectDefinitionEntity.getId()));
